+++ fabric/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [0.0.14-1.18.2] - 2022.09.01
+### Added
+- Added `activationStateWithoutEnchantment` configuration option, defaulted to `"KEYBINDING"`, which controls the
+  activation method if `maxBlocksBase` and `maxDistanceBase` values are greater than 0 (which would mean that the
+  enchantment is not required for vein mining functionality)
 ### Changed
 - `maxBlocksBase` and `maxBlocksPerLevel` now account for the origin block [#60](https://github.com/TheIllusiveC4/VeinMining/issues/60)